+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Drop support for phpdocumentor/reflection-docblock 4
 - Remove the unused symfony/deprecation-contracts dependency
 
+### Fixed
+
+- Lowercase the index prefix, as Elasticsearch rejects index names containing uppercase characters
+
 ## 2.1.0
 
 ### Added